Files refactored Manual and Automated Added state refactored to new file, statistics refactored to new file
