+++ automate.py
@@ -1,0 +1,6 @@
+import streamlit as st
+from state import session_state
+
+
+def automate():
+    st.title("Automated Page is Working ")
